ARM: dts: msm: Restore SDCC registers once Resuming from CX/XO
Defines the restore-after-cx-collapse property for MMC driver.
This basically lets the driver know that software needs to save the SDCC
register contents before the CX/XO event and restore the SDCC
register contents after coming out of CX/XO event.

Sdxlemur is not supporting retention of CX PC hence this change is
needed.

Change-Id: Id7808813509fb2ee0ce3469c6e04e787b466994b