ARM: dts: msm: enable pmu for 8295 vm
Enable arm pmu in gvm without pmu irq, because pass pmu ppi to
gvm is not supported by hyp yet.
This can provide basic pmu function for performance profiling,
the restriction is need to run perf with "-e clock-cpu" to use
hrtimer as interrupt for pmu events as overflow pmu irq is not
supported.

Change-Id: Ib9f58e4cc234820cfe3ad5ee5d7ea81f50283dd4