ARM: dts: msm: Add pin configuration for USB3 PHY portselect on direwolf
TLMM pins 166 and 49 are used to indicate the orientation in
which the type-C controller detected a USB-C connection, to the
USB3/DP Combo PHYs 0 and 1 respectively. Select this pinctrl
from the usb_qmp_dp_phy device and ensure it is selecting the
"usb0_phy" and "usb1_phy" pin function respectively.

Change-Id: I88c7f4be78751a0d3798889b8cfc1ecada94b654