ARM: dts: msm: Fix firmware-name in cdsp0 PIL node for Direwolf
subsys-pil-tz driver registers the subsystem with a name specified
in "qcom,firmware-name" property and later, firmware is looked up
based on the same name used as prefix.

For cdsp0, the firmware binaries are using prefix "cdsp", meaning,
the binaries are named as cdsp.mdt, cdsp.b00, cdsp.b01, and so on.
So, match the "qcom,firmware-name" property with the prefix used
for cdsp0 firmware.

Fixes: 5cb501598f8e ("ARM: dts: msm: Add subsys PIL node for CDSP0 on Direwolf")
Change-Id: Iea6b4f5515037091f9784237edca1e3619b3cffe