ARM: dts: msm: Enable vdd-1.8-xo only when server arrive
Currently vdd-1.8-xo is enabled during server arrive and
disabled after firmware ready indication. Again it is
enabled when wi-fi is on and remain in that state till wi-fi
is active which is drawing some leakage on this LDO.

Set regulator parameter "no_vote_on_wifi_active" for
vdd-1.8-xo which enables only during server arrive and
disables regulator after firmware ready indication.
In addition , this LDO is not voted on Wi-Fi ON.

Change-Id: I43bb119ce9eaa69a25595402f6147626d9d4870b