ARM: dts: msm: Add suspendable BW OPP table in SDMSHRIKE
We need to have the BW vote on behalf of NPU as long as
BW device is active. Replace the qcom,active-only flag with
the suspendable BW OPP table.

Also, add keepalive votes for Compute NoC and System NoC to
avoid these NoCs from being inactive while APPS is active.

Change-Id: I670f81dc27a3fd2b16e670ed2d9fd6767d75ca8c