ARM: dts: msm: Move apps_rsc earlier in devices for sa515m
Move the 'apps_rsc' node to the beginning of the 'soc'
node such that the rpmh-regulator devices are populated
early in the boot process. As many consumers' probe depend
on these regulator devices, placing 'apps_rsc' at the beginning
would satisfy the consumer devices as soon as they are
populated later in the boot. This way, the drivers no longer
have to wait for the regulator devices to be populated, thus
optimizing the boot performance.

While at this also move the pdc and intc interrupt controllers
at the beginning.

Change-Id: I0e54c83555d7d6ff596bfc3caab06f1a88ed6878