ARM: dts: qcom: Enable queue specific threshold mode
Enable queue specific threshold mode and control size so RX
queue controller will be enabled in cut-through mode.

Ethernet RX controller that uses cut-through mode can make a
forwarding decisionto DMA as soon as it gets the first couple
of bytes (32,64,96,128) of the incoming frame. The Rx Controller
does not have to wait for the rest of the frame to start switching
the frame from going out of the MTL block.

Change-Id: I8e150fbfa52f5fde0afdc52769e178f782d06c63