ARM: dts: msm: Correct the CX and MX voltage level from NOM to LOW_SVS
In sdxpinn,the max gen speed supported by PCIe RC1 & RC2 is GEN3 that
operates on LOW_SVS level.
But in the dt we vote the CX and MX voltage level to NOM for PCIe RC1
& RC2, which is not needed as these are not Gen4 capable.
As a result, we are seeing the CX and MX operating at the NOM corner
leading to higher power consumption.

Correct the CX and MX voltage level from NOM to LOW_SVS for sdxpinn.

Change-Id: Iddf1d4f66f51e9a7722e1a9782b9dbab2a7b6c21