ARM: dts: msm: Correct the CX and MX voltage level from NOM to LOW_SVS
In sdxbaagha, the max gen speed supported by PCIe RC is GEN2 that
operates on LOW_SVS level.
But in the dt we vote the CX and MX voltage level to NOM which is
required for GEN4. As a result we are seeing the CX and MX operating
at the NOM corner leading to higher power consumption.

Correct the CX and MX voltage level from NOM to LOW_SVS for sdxbaagha.

Change-Id: I79b6e2e5c1aa119ab1787ad2c0720593957f635c