ARM: dts: msm: Correct the MX voltage level from NOM to LOW_SVS
In sdxbaagha,the max gen speed supported by PCIe instance in endpoint
mode is GEN1 that operates on LOW_SVS level. But in the dt we vote the
MX voltage level to NOM for PCIe which is not needed as the controller
is not Gen4 capable. As a result, we are seeing the MX operating at the
NOM corner leading to higher power consumption.

Correct the MX voltage level from NOM to LOW_SVS for sdxbaagha.

Change-Id: I6547ecbea5032a1265ba8316bd5208a0a296ae37