ARM: dts: qcom: Disable halt feature at PCIe for sdxpinn
Change is to disable halt feature for sdxpinn at PCIe level, as
NOC already supports ordering feature. Enabling this at PCIe level
degrade the performance when bandwidth is above 9 Gbps.

This change is only enabled for WKK, as for HMT and ETH we dont
anticipate to cross the 9Gpbs bottleneck for these attaches.

Change-Id: Ib572b6a6a44e2a5f9fc6dc2d7788313000dabfd2